Adding 3.8.14-RC to trunk
git-svn-id: https://plugins.svn.wordpress.org/wp-e-commerce/trunk@905894 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ wpsc-admin/ajax.php
@@ -15,6 +15,7 @@
 function _wpsc_ajax_verify_nonce( $ajax_action ) {
 	// nonce can be passed with name wpsc_nonce or _wpnonce
 	$nonce = '';
+
 	if ( isset( $_REQUEST['nonce'] ) )
 		$nonce = $_REQUEST['nonce'];
 	elseif ( isset( $_REQUEST['_wpnonce'] ) )
@@ -191,26 +192,32 @@ function _wpsc_ajax_add_variation_set() {
 
 	require_once( 'includes/walker-variation-checklist.php' );
 
-	/* --- DIRTY HACK START --- */
-	/*
-	There's a bug with term cache in WordPress core. See http://core.trac.wordpress.org/ticket/14485.
-	The next 3 lines will delete children term cache for wpsc-variation.
-	Without this hack, the new child variations won't be displayed on "Variations" page and
-	also won't be displayed in wp_terms_checklist() call below.
-	*/
-	clean_term_cache( $variation_set_id, 'wpsc-variation' );
-	delete_option('wpsc-variation_children');
-	wp_cache_set( 'last_changed', 1, 'terms' );
-	_get_term_hierarchy('wpsc-variation');
-	/* --- DIRTY HACK END --- */
+	if ( ! version_compare( $GLOBALS['wp_version'], '3.8.3', '>' ) ) {
+
+		/* --- DIRTY HACK START --- */
+		/*
+		There's a bug with term cache in WordPress core. See http://core.trac.wordpress.org/ticket/14485. Fixed in 3.9.
+		The next 3 lines will delete children term cache for wpsc-variation.
+		Without this hack, the new child variations won't be displayed on "Variations" page and
+		also won't be displayed in wp_terms_checklist() call below.
+		*/
+		clean_term_cache( $variation_set_id, 'wpsc-variation' );
+		delete_option('wpsc-variation_children');
+		wp_cache_set( 'last_changed', 1, 'terms' );
+		_get_term_hierarchy('wpsc-variation');
+		/* --- DIRTY HACK END --- */
+
+	}
 
 	ob_start();
+
 	wp_terms_checklist( (int) $_POST['post_id'], array(
 		'taxonomy'      => 'wpsc-variation',
 		'descendants_and_self' => $variation_set_id,
 		'walker'        => new WPSC_Walker_Variation_Checklist( $inserted_variants ),
 		'checked_ontop' => false,
 	) );
+
 	$content = ob_get_clean();
 
 	$return = array(
@@ -380,8 +387,7 @@ function _wpsc_ajax_purchase_log_send_tracking_email() {
 	$message = str_replace( '%trackid%', $trackingid, $message );
 	$message = str_replace( '%shop_name%', get_option( 'blogname' ), $message );
 
-	$email_form_field = $wpdb->get_var( "SELECT `id` FROM `" . WPSC_TABLE_CHECKOUT_FORMS . "` WHERE `type` IN ('email') AND `active` = '1' ORDER BY `checkout_order` ASC LIMIT 1" );
-	$email = $wpdb->get_var( $wpdb->prepare( "SELECT `value` FROM `" . WPSC_TABLE_SUBMITTED_FORM_DATA . "` WHERE `log_id`=%d AND `form_id` = '$email_form_field' LIMIT 1", $id ) );
+	$email = wpsc_get_buyers_email( $id );
 
 	$subject = get_option( 'wpsc_trackingid_subject' );
 	$subject = str_replace( '%shop_name%', get_option( 'blogname' ), $subject );
@@ -460,8 +466,6 @@ function _wpsc_ajax_remove_product_meta() {
  *
  * @uses wpsc_purchlog_edit_status()                    Edits purchase log status
  * @uses WP_Error                                       WordPress Error class
- * @uses get_bloginfo()                                 Gets information about your WordPress site
- * @uses set_current_screen()                           Sets current screen object
  * @uses WPSC_Purchase_Log_List_Table
  * @uses WPSC_Purchase_Log_List_Table::prepare_items()
  * @uses WPSC_Purchase_Log_List_Table::views()
@@ -476,10 +480,7 @@ function _wpsc_ajax_change_purchase_log_status() {
 
 	$args = array();
 
-	if ( version_compare( get_bloginfo( 'version' ), '3.5', '<' ) )
-		set_current_screen( 'dashboard_page_wpsc-sales-logs' );
-	else
-		$args['screen'] = 'dashboard_page_wpsc-sales-logs';
+	$args['screen'] = 'dashboard_page_wpsc-sales-logs';
 
 	require_once( WPSC_FILE_PATH . '/wpsc-admin/includes/purchase-log-list-table-class.php' );
 	$purchaselog_table = new WPSC_Purchase_Log_List_Table( $args );
